Проч сран

+++ app/src/main/java/com/example/mall_navigation/models/RouteRequest.kt
@@ -1,7 +1,7 @@
 package com.example.mall_navigation.models
 
 class RouteRequest (
-    val floor: Int,
+    val floor: Double,
     val endX: Double,
     val startX: Double,
     val startY: Double,
